+++ assets/js/components/index.js
@@ -1,7 +1,6 @@
 // common (reusable) components are listed here
 
 export { default as Flash } from "./flash"
-export { default as Footer } from "./footer"
 export { default as Header } from "./header"
 export { default as Icon } from "./icon"
 export { default as Loading } from "./loading"